Update README.md
This is our first project, and it is the first practice of designing a web page, in this project me and my group members make a wireframe and mock-up for a web site under title Miracle Hands. On our web site we offer help to people working in handicrafts to sell their products, and we've designed 9 pages: Home, log in, sign in, Bags, Single product, cart, Payment, about us and contact us.

I designed a wireframe and mock-up for Cart, Payment and contact us pages.